Changed the Feature and Object classes so that Object inherited from Feature in order to prevent repeated code.
