+++ feature.py
@@ -1,5 +1,4 @@
 from interactable import Interactable
-from object import Object
 
 class FeatureException(Exception):
     """Base exception class for the Feature class"""
@@ -13,10 +12,31 @@ class AttributeAlreadyExists(FeatureException):
     """Raised when attribute already exists"""
     pass
 
-class Feature(Object):
+class Feature(Interactable):
 
-    def __init__(self, id: str, name: str, desc: str, sdesc: str, interactions: dict, hidden: bool):
-        super().__init__(id, name, desc, sdesc, interactions, hidden)
+    def __init__(self, id: str, name: str, desc: str, sdesc: str, interactions: dict, 
+    hidden: bool):
+        super().__init__(id, name, desc, sdesc, interactions)
+        # the visible/hidden status of the object
+        self._hidden = hidden
+
+    # methods for managing object
+    def get_hidden(self):
+        """returns the value of self._hidden"""
+        return self._hidden
+
+    def set_hidden(self, hidden: bool):
+        """changes the value of self._hidden to the value provided in
+        the hidden argument"""
+        self._hidden = hidden
+
+    def switch_hidden(self):
+        """changes the value of self._hidden to the opposite of its
+        current Boolean value"""
+        if self._hidden:
+            self._hidden = False
+        else:
+            self._hidden = True
 
     # methods for managing the optional condition variable
     def add_condition(self, condition: bool):
